pyproject.toml
[tool.poetry]
name = "deepsearch-examples"
version = "0.2.0"
description = "Showcase some common usage of Deep Search for Document conversion as well as Data and Knowledge exploration."
authors = ["Michele Dolfi <dol@zurich.ibm.com>"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
deepsearch-toolkit = "^0.3.4"
jupyter = "^1.0.0"
ipywidgets = "^8.0.2"
numpy = "^1.23.4"
matplotlib = "^3.6.2"
pillow = "^9.3.0"
pandas = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = """
# Exclude generated API code
^/deepsearch/cps/apis/.+$
"""

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["docs", "deepsearch/cps/apis"]
py_version=38
known_first_party = ["cps"]

[tool.mypy]
pretty = true
# strict = true
no_implicit_optional = true
python_version = 3.8

# We are not going to bother about generated code...
[[tool.mypy.overrides]]
module = "deepsearch.cps.apis.*"
ignore_errors = true
# https://stackoverflow.com/a/60744115
implicit_reexport = true

# Silence errors for third-party packages that don't have
# typings available.
[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "IPython.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ipywidgets.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ipython_blocking.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rdkit.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mhfp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "appdirs.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pluggy.*"
ignore_missing_imports = true

